+++ bundle.sh
@@ -1,0 +1,5 @@
+set -ex
+
+# Bundle
+mkdir -p dist
+deno bundle color-hash.ts dist/color-hash.js
