+++ lib/bkdr-hash.ts
@@ -4,21 +4,21 @@
  * @param {String} str string to hash
  * @returns {Number}
  */
-var BKDRHash = function(str) {
+const BKDRHash = function(str: string) {
     var seed = 131;
     var seed2 = 137;
     var hash = 0;
     // make hash more sensitive for short string like 'a', 'b', 'c'
     str += 'x';
     // Note: Number.MAX_SAFE_INTEGER equals 9007199254740991
-    var MAX_SAFE_INTEGER = parseInt(9007199254740991 / seed2);
+    var MAX_SAFE_INTEGER = Math.floor(9007199254740991 / seed2);
     for(var i = 0; i < str.length; i++) {
         if(hash > MAX_SAFE_INTEGER) {
-            hash = parseInt(hash / seed2);
+            hash = Math.floor(hash / seed2);
         }
         hash = hash * seed + str.charCodeAt(i);
     }
     return hash;
 };
 
-module.exports = BKDRHash;
+export default BKDRHash;
